fix: suppress C4244 warning on Lua::ToNumber narrowing conversion
ToNumber returns double as required by the Lua API. The autoloot value
is always 0 or 1 so there is no actual data loss. Add an explicit
static_cast<int> to document the intentional narrowing and silence the
warning.

+++ src/dllmain.cpp
@@ -131,7 +131,7 @@ static uint32_t InteractNearest(void *L) {
   if (!finalCandidate)
     return 0;
 
-  int autoloot = Lua::ToNumber(L, 1);
+  int autoloot = static_cast<int>(Lua::ToNumber(L, 1));
 
   if (finalCandidate->type == ObjectType::UNIT) {
     Game::SetTarget(finalCandidate->guid);
